Initial files
These are some files I coded up to scrape the data to get started. ncaa.db should contain all the data so you shouldn't have to even touch march_madness_scrape, but I would love for people to look through it and offer suggestions of how to change it to make it more efficient.